% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EW_score.R
\name{EW_score}
\alias{EW_score}
\title{Score each decision tree}
\usage{
EW_score(data)
}
\arguments{
\item{data}{Data is a dataframe with at least 11 columns including names of miRNAs, k, i, frequency of miRNAs, the value of single ROC, combined ROC, prerat, sensitivity, specificity, precision and F_measure of evaluate set and test set.K and i are the ordinal number of the feature selection set.}
}
\value{
The value of score of every decision tree.
}
\description{
Score each decision tree
}
\examples{
EW_score(data)

}
